+++ scripts/mruby/darwin.rb
@@ -20,6 +20,7 @@ MRuby::CrossBuild.new('x86_64-apple-darwin19') do |conf|
   conf.gembox "metaprog"
 
   conf.gem core: 'mruby-exit'
+  conf.gem core: 'mruby-sleep'
   conf.gem github: 'iij/mruby-dir'
   conf.gem github: 'iij/mruby-env'
   conf.gem github: 'iij/mruby-iijson'