+++ scripts/mruby/emscripten.rb
@@ -14,6 +14,7 @@ MRuby::CrossBuild.new('emscripten') do |conf|
   conf.gembox "metaprog"
 
   conf.gem core: 'mruby-exit'
+  conf.gem core: 'mruby-sleep'
   conf.gem github: 'iij/mruby-dir'
   conf.gem github: 'iij/mruby-env'
   conf.gem github: 'iij/mruby-iijson'