+++ scripts/mruby/linux.rb
@@ -9,9 +9,10 @@ MRuby::Build.new do |conf|
   conf.gembox "metaprog"
 
   # Generate mrbc command
-  conf.gem :core => "mruby-bin-mrbc"
+  conf.gem core: "mruby-bin-mrbc"
 
   conf.gem core: 'mruby-exit'
+  conf.gem core: 'mruby-sleep'
   conf.gem github: 'iij/mruby-dir'
   conf.gem github: 'iij/mruby-env'
   conf.gem github: 'iij/mruby-iijson'