+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- mRuby now builds from master instead of 3.0.0
+- Add mruby-sleep as a library
 
 ## v0.3.4